@@ -9,8 +9,8 @@
   "dependencies": {
     "@heroicons/react": "^1.0.6",
     "clsx": "^1.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-query": "^3.34.19",
     "react-router-dom": "^6.3.0"
   },
